requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
Django==3.2
django-keycloak-auth==0.9.2
djangorestframework==3.12.4
idna==2.10
josepy==1.8.0
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
